services/scraping/requirements.txt
python-dotenv
requests
aiohttp